Use custom connection to return non-null edges
