+++ saleor/graphql/core/connection.py
@@ -1,0 +1,42 @@
+import graphene
+from graphene import Field, ObjectType, String, NonNull, List
+from graphene.relay.connection import ConnectionOptions, Connection
+
+
+class NonNullConnection(Connection):
+
+    class Meta:
+        abstract = True
+
+    @classmethod
+    def __init_subclass_with_meta__(cls, node=None, name=None, **options):
+        super().__init_subclass_with_meta__(node=node, name=name, **options)
+
+        # Override the original EdgeBase type to make to `node`` field required.
+        class EdgeBase(object):
+            node = Field(
+                cls._meta.node, description='The item at the end of the edge',
+                required=True)
+            cursor = String(
+                required=True, description='A cursor for use in pagination')
+
+        # Create the edge type using the new EdgeBase.
+        edge_name = cls.Edge._meta.name
+        edge_bases = (EdgeBase, ObjectType,)
+        edge = type(edge_name, edge_bases, {})
+        cls.Edge = edge
+
+        # Override the `edges` field to make it non-null list of non-null edges.
+        cls._meta.fields['edges'] = Field(NonNull(List(NonNull(cls.Edge))))
+
+
+class CountableConnection(NonNullConnection):
+    class Meta:
+        abstract = True
+
+    total_count = graphene.Int(
+        description='A total count of items in the collection')
+
+    @staticmethod
+    def resolve_total_count(root, info, *args, **kwargs):
+        return root.length
